// src/app/models/admin/capphephoatdongkhoangsan/cppheduyettruluongkhoangsan/cpPheDuyetTruLuongKhoiTruLuong.model.ts
export class OutputCpPheDuyetTruLuongKhoiTruLuongModel {
    idpheduyettruluongkhoitruluong: string;
    idcaptruluong: string;
    tenkhoitruluong: string;
    truluong: number;
    mucsaucaodothapnhat: number;
    mucsaucaodocaonhat: number;
    donvitruluong: string;
    donvichieusau: string;
    thutu: number;
    ghichu: string;
    idpheduyettruluong: string
  }
  
  export class InputCpPheDuyetTruLuongKhoiTruLuongModel {
    idpheduyettruluongkhoitruluong: string;
    idcaptruluong: string;
    tenkhoitruluong: string;
    truluong: number;
    mucsaucaodothapnhat: number;
    mucsaucaodocaonhat: number;
    donvitruluong: string;
    donvichieusau: string;
    thutu: number;
    ghichu: string;
    idpheduyettruluong: string
  }
  